@@ -1,19 +1,13 @@
 package com.baeldung.spring.data.es.config;
 
-import java.net.InetAddress;
-import java.net.UnknownHostException;
-
-import org.elasticsearch.client.Client;
-import org.elasticsearch.client.transport.TransportClient;
-import org.elasticsearch.common.settings.Settings;
-import org.elasticsearch.common.transport.InetSocketTransportAddress;
-import org.elasticsearch.transport.client.PreBuiltTransportClient;
-import org.springframework.beans.factory.annotation.Value;
+import org.elasticsearch.client.RestHighLevelClient;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.data.elasticsearch.client.ClientConfiguration;
+import org.springframework.data.elasticsearch.client.RestClients;
 import org.springframework.data.elasticsearch.core.ElasticsearchOperations;
-import org.springframework.data.elasticsearch.core.ElasticsearchTemplate;
+import org.springframework.data.elasticsearch.core.ElasticsearchRestTemplate;
 import org.springframework.data.elasticsearch.repository.config.EnableElasticsearchRepositories;
 
 @Configuration
@@ -21,30 +15,18 @@ import org.springframework.data.elasticsearch.repository.config.EnableElasticsea
 @ComponentScan(basePackages = { "com.baeldung.spring.data.es.service" })
 public class Config {
 
-    @Value("${elasticsearch.home:/usr/local/Cellar/elasticsearch/5.6.0}")
-    private String elasticsearchHome;
-
-    @Value("${elasticsearch.cluster.name:elasticsearch}")
-    private String clusterName;
-
     @Bean
-    public Client client() {
-        TransportClient client = null;
-        try {
-            final Settings elasticsearchSettings = Settings.builder()
-              .put("client.transport.sniff", true)
-              .put("path.home", elasticsearchHome)
-              .put("cluster.name", clusterName).build();
-            client = new PreBuiltTransportClient(elasticsearchSettings);
-            client.addTransportAddress(new InetSocketTransportAddress(InetAddress.getByName("127.0.0.1"), 9300));
-        } catch (UnknownHostException e) {
-            e.printStackTrace();
-        }
-        return client;
+    RestHighLevelClient client() {
+        ClientConfiguration clientConfiguration = ClientConfiguration.builder()
+            .connectedTo("localhost:9200")
+            .build();
+
+        return RestClients.create(clientConfiguration)
+            .rest();
     }
 
     @Bean
     public ElasticsearchOperations elasticsearchTemplate() {
-        return new ElasticsearchTemplate(client());
+        return new ElasticsearchRestTemplate(client());
     }
 }
