@@ -8,10 +8,10 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import com.baeldung.spring.data.es.config.Config;
 
 /**
+ * This Manual test requires: Elasticsearch instance running on localhost:9200.
  * 
- * This Manual test requires:
- * * Elasticsearch instance running on host
- *
+ * The following docker command can be used: docker run -d --name es762 -p
+ * 9200:9200 -e "discovery.type=single-node" elasticsearch:7.6.2
  */
 @RunWith(SpringJUnit4ClassRunner.class)
 @ContextConfiguration(classes = Config.class)
